+++ homeassistant/components/zha/__init__.py
@@ -61,6 +61,7 @@ from .repairs.wrong_silabs_firmware import (
     AlreadyRunningEZSP,
     warn_on_wrong_silabs_firmware,
 )
+from .services import async_setup_services
 
 DEVICE_CONFIG_SCHEMA_ENTRY = vol.Schema({vol.Optional(CONF_TYPE): cv.string})
 ZHA_CONFIG_SCHEMA = {
@@ -121,6 +122,8 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
     hass.data[DATA_ZHA] = ha_zha_data
 
     async_register_firmware_info_provider(hass, DOMAIN, homeassistant_hardware)
+
+    async_setup_services(hass)
 
     return True
 
@@ -298,8 +301,6 @@ async def async_unload_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> 
     with contextlib.suppress(KeyError):
         for platform in PLATFORMS:
             del ha_zha_data.platforms[platform]
-
-    websocket_api.async_unload_api(hass)
 
     return True
