+++ homeassistant/components/zha/helpers.py
@@ -60,6 +60,8 @@ from zha.application.helpers import (
     QuirksConfiguration,
     ZHAConfiguration,
     ZHAData,
+    convert_install_code,
+    qr_to_install_code,
 )
 from zha.application.platforms import GroupEntity, PlatformEntity
 from zha.event import EventBase
@@ -118,27 +120,31 @@ from homeassistant.helpers import (
 )
 from homeassistant.helpers.dispatcher import async_dispatcher_send, dispatcher_send
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.typing import ConfigType
+from homeassistant.helpers.typing import ConfigType, VolDictType
 from homeassistant.util.logging import HomeAssistantQueueHandler
 
 from .const import (
     ATTR_ACTIVE_COORDINATOR,
     ATTR_AVAILABLE,
     ATTR_DEVICE_TYPE,
+    ATTR_DURATION,
     ATTR_ENDPOINT_NAMES,
     ATTR_EXPOSES_FEATURES,
     ATTR_IEEE,
+    ATTR_INSTALL_CODE,
     ATTR_LAST_SEEN,
     ATTR_LQI,
     ATTR_MANUFACTURER_CODE,
     ATTR_NEIGHBORS,
     ATTR_NWK,
     ATTR_POWER_SOURCE,
+    ATTR_QR_CODE,
     ATTR_QUIRK_APPLIED,
     ATTR_QUIRK_CLASS,
     ATTR_ROUTES,
     ATTR_RSSI,
     ATTR_SIGNATURE,
+    ATTR_SOURCE_IEEE,
     CONF_ALARM_ARM_REQUIRES_CODE,
     CONF_ALARM_FAILED_TRIES,
     CONF_ALARM_MASTER_CODE,
@@ -1466,3 +1472,18 @@ def exclude_none_values(obj: Mapping[str, Any]) -> dict[str, Any]:
 def get_config_entry_unique_id(network_info: NetworkInfo) -> str:
     """Generate a unique id for a config entry based on the network info."""
     return f"epid={network_info.extended_pan_id}".lower()
+
+
+IEEE_SCHEMA = vol.All(cv.string, EUI64.convert)
+
+SERVICE_PERMIT_PARAMS: VolDictType = {
+    vol.Optional(ATTR_IEEE): IEEE_SCHEMA,
+    vol.Optional(ATTR_DURATION, default=60): vol.All(
+        vol.Coerce(int), vol.Range(0, 254)
+    ),
+    vol.Inclusive(ATTR_SOURCE_IEEE, "install_code"): IEEE_SCHEMA,
+    vol.Inclusive(ATTR_INSTALL_CODE, "install_code"): vol.All(
+        cv.string, convert_install_code
+    ),
+    vol.Exclusive(ATTR_QR_CODE, "install_code"): vol.All(cv.string, qr_to_install_code),
+}
